Split reader module
